requirements.txt
numpy
opencv-python
pyyaml
scipy
matplotlib
torch
tqdm
